Pre-testing fix
