+++ src/passes/references/storedPointerDereference.ts
@@ -1,22 +1,20 @@
 import assert from 'assert';
 import {
-  ArrayType,
   Assignment,
   DataLocation,
   Expression,
   FunctionCall,
-  generalizeType,
   getNodeType,
   IndexAccess,
-  MappingType,
   MemberAccess,
-  PointerType,
-  StructDefinition,
-  TypeNode,
-  UserDefinedType,
 } from 'solc-typed-ast';
 import { AST } from '../../ast/ast';
-import { typeNameFromTypeNode } from '../../utils/utils';
+import {
+  isComplexMemoryType,
+  isDynamicStorageArray,
+  isMapping,
+  typeNameFromTypeNode,
+} from '../../utils/utils';
 import { ReferenceSubPass } from './referenceSubPass';
 
 export class StoredPointerDereference extends ReferenceSubPass {
@@ -72,27 +70,4 @@ export class StoredPointerDereference extends ReferenceSubPass {
       this.visitExpression(node, ast);
     }
   }
-}
-
-function isDynamicStorageArray(type: TypeNode): boolean {
-  return (
-    type instanceof PointerType &&
-    type.location === DataLocation.Storage &&
-    type.to instanceof ArrayType &&
-    type.to.size === undefined
-  );
-}
-
-function isComplexMemoryType(type: TypeNode): boolean {
-  return (
-    type instanceof PointerType &&
-    type.location === DataLocation.Memory &&
-    (type.to instanceof ArrayType ||
-      (type.to instanceof UserDefinedType && type.to.definition instanceof StructDefinition))
-  );
-}
-
-function isMapping(type: TypeNode): boolean {
-  const [base] = generalizeType(type);
-  return base instanceof MappingType;
 }
